Revert "Merge branch 'master' of https://github.com/Suprcode/mir2"
This reverts commit cf0721f09254b96f39e3163d96c9a70d304c14d6, reversing
changes made to 8baa30f415761e0e1e710a8fca2e1c9ca969f1b7.

+++ Client/MirObjects/DecoObject.cs
@@ -21,7 +21,7 @@ namespace Client.MirObjects
             get { return false; }
         }
 
-        public int Image;
+        public ushort Image;
 
         public DecoObject(uint objectID)
             : base(objectID)
